fix: move each class related to upload propagation to their own files
